Add files via upload
In the era of technological innovation, smartphones have become an integral part of our daily lives. The iPhone, a flagship product from Apple Inc., stands as a symbol of cutting-edge technology and design. This project delves into a comprehensive analysis of iPhone data using Python and various libraries, aiming to unveil insightful patterns and trends in the realm of mobile devices.